refactor(reposicao): split AdminReposicaoSlaFornecedor (576→104) em hook + componentes
Extrai verbatim de src/pages/AdminReposicaoSlaFornecedor.tsx:
- useSlaFornecedor: 3 queries (compliance por fornecedor/SKU + histórico)
  + memos (grupos/skusFiltrados/fornecedoresOptions) + exportCsv + toggleStatus
- ComplianceCardsGrid, SlaFiltros, StatusChips, SkuComplianceTable,
  SkuHistoricoDialog, TendenciaIcon, types.ts, config.ts
  (STATUS_LABEL/VARIANT/RANK + fmtNum/fmtData/desvioColorClass/cardTone)
- +20 testes (helpers puros + presentacionais)

Comportamento preservado 1:1. tsc baseline=0, eslint=0.
